chore: rename file and add images
