+++ src/components/LinkButton.tsx
@@ -36,16 +36,18 @@ const styles = {
     ],
 };
 
-const LinkButton = ({
+function LinkButton({
     color = 'violet-400', size = 'medium', children = 'Button', href = '/', ...props
-}: LinkButtonProps) => (
-    <StyledLinkButton
-        css={styles.style({ color, size, children })}
-        href={href}
-        {...props}
-    >
-        {children}
-    </StyledLinkButton>
-);
+}: LinkButtonProps) {
+    return (
+        <StyledLinkButton
+            css={styles.style({ color, size, children })}
+            href={href}
+            {...props}
+        >
+            {children}
+        </StyledLinkButton>
+    );
+}
 
 export default LinkButton;
